Moved the log module
